Added comment to ui
