+++ src/index.js
@@ -1,5 +1,4 @@
 import './scss/main.css';
-import  getData  from './module/CardApi.js';
+import getData from './module/CardApi.js';
 
 getData();
-
